backend/src/requirements.txt
bleach==6.0.0
bottle==0.12.25
bottle-cors==0.1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cygrpc==1.0.4
Cython==3.0.0b2
docutils==0.19
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-documentai==2.15.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.1
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
grpcio-tools==1.53.0
idna==3.4
importlib-metadata==6.4.1
jaraco.classes==3.2.3
keyring==23.13.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
Paste==3.5.2
pkginfo==1.9.6
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.4
rsa==4.9
six==1.16.0
twine==4.0.2
urllib3==1.26.15
webencodings==0.5.1
zipp==3.15.0
